requirements.txt
Flask
redis
jieba
bson
pandas
opencv-python
pillow
torch
torchvision
gevent-websocket
requests
efficientnet_pytorch
scikit-image